pyproject.toml
[project]
name = "sa"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "hydra-core",
    "mlflow",
    "torch",
    "scikit-learn",
    "pandas",
    "uv",
    "torchbnn",
]
